added profiling info; timer tweaks

+++ src/util.h
@@ -8,8 +8,8 @@ namespace util {
 
 // start new time measurement
 void start_timer();
-// returns time in milliseconds since last unconsumed start_timer call and consumes it
-long stop_timer();
+// returns time in seconds since last unconsumed start_timer call and consumes it
+double stop_timer();
 
 // sort vector and remove duplicate elements
 template<typename T>
